fixed divide method

+++ src/com/aut/AUTMath.java
@@ -2,24 +2,24 @@ package com.aut;
 
 public class AUTMath {
     /**
-     * gets two integers and returns their sum.
+     * gets two doubles and returns their sum.
      *
      * @param num1 the num 1
      * @param num2 the num 2
-     * @return int
+     * @return double
      */
-    public static int sum(int num1, int num2) {
+    public static double sum(double num1, double num2) {
         return num1+num2;
     }
 
     /**
-     * gets two integer and returns their subtraction.
+     * gets two doubles and returns their subtraction.
      *
      * @param num1 the num 1
      * @param num2 the num 2
-     * @return int
+     * @return double
      */
-    public static int subtract(int num1, int num2) {
+    public static double subtract(double num1, double num2) {
         return num1 - num2;
     }
 
@@ -28,9 +28,9 @@ public class AUTMath {
      *
      * @param num1 the num 1
      * @param num2 the num 2
-     * @return int
+     * @return double
      */
-    public static int multiply(int num1, int num2) {
+    public static double multiply(double num1, double num2) {
         return num1 * num2;
     }
 
@@ -39,9 +39,9 @@ public class AUTMath {
      *
      * @param num1 the num 1
      * @param num2 the num 2
-     * @return int
+     * @return double
      */
-    public static int divide(int num1, int num2) {
+    public static double divide(double num1, double num2) {
         return num1 / num2;
     }
 
@@ -49,10 +49,10 @@ public class AUTMath {
      * gets a number and returns its factorial.
      *
      * @param number the number
-     * @return int
+     * @return double
      */
-    public static int factorial(int number) {
-        int result = 0;
+    public static double factorial(double number) {
+        double result = 0;
 
         for (int counter = 0; counter < number; counter++) {
             result = multiply(result, number);
@@ -65,10 +65,10 @@ public class AUTMath {
      *
      * @param base  the base
      * @param power the power
-     * @return int
+     * @return double
      */
-    public static int pow(int base, int power) {
-        int result = 1;
+    public static double pow(double base, double power) {
+        double result = 1;
 
         for (int counter = 1; counter <= base; counter++) {
             result = multiply(result, power);
@@ -81,9 +81,9 @@ public class AUTMath {
      * gets an index and calculates the fibonacci sequence at that exact index recursively.
      *
      * @param index the index
-     * @return int
+     * @return double
      */
-    public static int fib(int index) {
+    public static double fib(int index) {
         if (index == 0 || index == 1) {
             return index;
         }
